0324/requirements.txt
langchain-core>=0.3.0
langchain-openai>=0.3.0
python-dotenv

